refactor: extract `RecentAlgorithm` from old room list store
Move `RecentAlgorithm` and `Algorithm`  to `utils/room-sorting`. They
are used in some place of the code to do a manual sorting of the rooms.
