+++ apps/web/src/stores/room-list/algorithms/tag-sorting/AlphabeticAlgorithm.ts
@@ -9,12 +9,12 @@ Please see LICENSE files in the repository root for full details.
 import { type Room } from "matrix-js-sdk/src/matrix";
 
 import { type TagID } from "../../../room-list-v3/skip-list/tag";
-import { type IAlgorithm } from "./IAlgorithm";
+import { type Algorithm } from "../../../../utils/room/sorting/Algorithm";
 
 /**
  * Sorts rooms according to the browser's determination of alphabetic.
  */
-export class AlphabeticAlgorithm implements IAlgorithm {
+export class AlphabeticAlgorithm implements Algorithm {
     public sortRooms(rooms: Room[], tagId: TagID): Room[] {
         const collator = new Intl.Collator();
         return rooms.sort((a, b) => {
